+++ Shop.Services/ModelServices/TypeService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
 using Shop.Data.ViewModels;
@@ -24,14 +25,24 @@ namespace Shop.Services.ModelServices
             return await _repository.GetAll();
         }
 
-        public async Task<Type> GetById(long Id)
+        public async Task<Type> GetById(long id)
         {
-            return await _repository.GetById(Id);
+            var type = await _repository.GetById(id);
+            if (type == null)
+            {
+                throw new Exception($"Cannot find Type with id: {id}");
+            }
+            return type;
         }
 
         public async Task<Type> Add(TypeVM vm)
         {
             var type = _mapper.Map<Type>(vm);
+            var matched = await _repository.GetSameType(type);
+            if (matched != null)
+            {
+                throw new Exception("Same Type already exists");
+            }
             if (type == null)
             {
                 return null;
@@ -41,17 +52,31 @@ namespace Shop.Services.ModelServices
 
         public async Task<Type> Update(TypeVM vm, long id)
         {
-            var type = await GetById(id);
+            var type = await _repository.GetById(id);
+            if (type == null)
+            {
+                throw new Exception($"Cannot find Type with id: {id}");
+            }
             if (vm == null)
             {
                 return null;
             }
             type.Name = vm.Name;
+            var matched = await _repository.GetSameType(type);
+            if (matched != null)
+            {
+                throw new Exception("Same Type already exists");
+            }
             return await _repository.Update(type);
         }
 
         public async Task Delete(long id)
         {
+            var matched = await _repository.GetById(id);
+            if (matched == null)
+            {
+                throw new Exception($"Cannot find Type with id: {id}");
+            }
             await _repository.Delete(id);
         }
     }
